// src/app/models/inspeccion.ts
export interface Inspeccion {

    IdInspeccion: number;
    Nombre: string;
    Apellido: string;
    DocumentoIdentificacion: string;
    Direccion: string;
    Telefono: string;
    Municipio: string;
    Producto: string;
    PlantaElectrica: string;
    TipoVehiculo: string;
    CondicionesTerreno: string;
    ConexionesElectricas: string;
    TecnicoRequerido: string;
    HerramientasEquipos: string;
    Largo: string;
    Ancho: string;
    Fondo: string;
    Diametro: string;
    Capacidad: string;
    Material: string;
    Cantidad: string;
    TecnicoRealizaInspeccion: string;
    PersonaAcompanaInspeccion: string;
    Observaciones: string;
}